+++ source/Examples/Resources/xSccmSqlSetup/SccmSqlSetup.ps1
@@ -1,0 +1,36 @@
+#Requires -Module ConfigMgrCBDsc
+
+<#
+    .DESCRIPTION
+        This configuration Install SQL for an SCCM install. The DSC resource module, SqlServerDsc, can be used to
+        create a more sophisticated install.
+#>
+Configuration Example
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $SqlServiceCredential,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $SqlAgentServiceCredential
+    )
+
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        xSccmSqlSetup SccmSqlSetup
+        {
+            SqlVersion                = '2014'
+            SqlInstanceName           = 'CASINST01'
+            SqlServiceCredential      = $SqlServiceCredential
+            SqlAgentServiceCredential = $SqlAgentServiceCredential
+            SqlSysAdminAccounts       = @('contoso\SqlAdmin01','contoso\SqlAdmin02','contoso\SqlAdminGroup')
+            SqlInstallPath            = 'C:\Windows\Temp\SQL\MSSQL2014wSP3'
+        }
+    }
+}